requirements.txt
flask==3.0.3
pycryptodomex==3.20.0